Move action typings to separate file
